Add files via upload
This project was completed by insert full name here in partial fulfilment of ECON-UB.0232,
Data Bootcamp, Spring 2019. I certify that the NYU Stern Honor Code applies to this project.
In particular, I have:
Clearly acknowledged the work and efforts of others when submitting written work as our own.
The incorporation of the work of others–including but not limited to their ideas, data, creative
expression, and direct quotations (which should be designated with quotation marks), or paraphrasing thereof– has been fully and appropriately referenced using notations both in the text
and the bibliography.
And I understand that:
Submitting the same or substantially similar work in multiple courses, either in the same semester
or in a different semester, without the express approval of all instructors is strictly forbidden.
I acknowledge that a failure to abide by NYU Stern Honor Code will result in a failing grade for
the project and course.
Project Description

My project investigates the relationship between the Female Labor Force Participation Rate and GDP. I do so by graphing individual countries that perform well or poorly in terms of FLFPR and GDP and compare them. I graph the top percentiles in terms of FLFPR growth on bar charts that measure how each percentile performs in terms of FLFPR growth and GDP growth. Lastly, I create a scatterplot that proves a positive association between the two factors exists.